pyproject.toml
[project]
name = "gradio-gemini"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.39.1",
    "gradio>=5.47.2",
    "python-dotenv>=1.1.1",
]
